Add check for full site links and refactor tooltip messages
This commit introduces a new check that triggers a manual review if full links are used within the same site. This is meant to discourage the use of Fully Qualified Domain Names (FQDN) for internal links. Refactoring was also done to re-arrange the tooltip messages, suggesting the use of relative links instead of full URL or dev links wherever possible.