Add option for default 404 page in Madrone theme
An option to use a theme-provided 404 page has been included in the Madrone theme settings. This aims to give users more flexibility in page not found scenarios. A checkbox has been added to the theme settings form, and corresponding configurations and conditional rendering in the Twig file have been established to implement this feature.