Update play/pause button icon in video controls
The icons in the play/pause buttons of the video controls have been updated to use Font Awesome. Before, the text element in the button was directly manipulated, but now it has been replaced with an HTML string that includes a Font Awesome icon. This change has been reflected in both the .js and .twig files to maintain consistency.