Add styles and twig template for media library widget modal
This commit introduces a number of new CSS rules to style the media library widget modal. Additionally, a new Twig template has been added to theme the media items shown in the grid of the modal. The changes aim to improve the display and user interaction with the media library widget modal.