Updated level 1 heading guidance in accessibility tips
The guidance messages for the missing level 1 heading in both the source and distribution files have been updated. The new messages provide more detailed instructions on how to add a level 1 heading using the Layout Builder. Also, the guidelines now handle situations where the Page Title may not serve as the preferred level 1 heading for the user.