Refactor site branding system block
Extracted site link as a macro in the 'block--system-branding-block.html.twig' template. Also added conditions to handle the existence of parent site name and URL in the system branding block. Adjusted the handling of head title and footer in the 'madrone.theme' file, including the additions of parent site name and URL.